+++ minicbor-derive/src/decode.rs
@@ -66,9 +66,9 @@ fn on_struct(inp: &mut syn::DeriveInput) -> syn::Result<proc_macro2::TokenStream
     let params = inp.generics.type_params_mut();
     add_bound_to_matching_type_params(bound, params, &default_types);
 
-    let gen = add_lifetime(&inp.generics, lifetime);
-    let gen = add_typeparam(&gen, gen_ctx_param()?, attrs.context_bound());
-    let impl_generics = gen.split_for_impl().0;
+    let generics = add_lifetime(&inp.generics, lifetime);
+    let generics = add_typeparam(&generics, gen_ctx_param()?, attrs.context_bound());
+    let impl_generics = generics.split_for_impl().0;
 
     let (_, typ_generics, where_clause) = inp.generics.split_for_impl();
 
@@ -235,9 +235,9 @@ fn on_enum(inp: &mut syn::DeriveInput) -> syn::Result<proc_macro2::TokenStream> 
     let params = inp.generics.type_params_mut();
     add_bound_to_matching_type_params(bound, params, &defaults);
 
-    let gen = add_lifetime(&inp.generics, lifetime);
-    let gen = add_typeparam(&gen, gen_ctx_param()?, enum_attrs.context_bound());
-    let impl_generics = gen.split_for_impl().0;
+    let generics = add_lifetime(&inp.generics, lifetime);
+    let generics = add_typeparam(&generics, gen_ctx_param()?, enum_attrs.context_bound());
+    let impl_generics = generics.split_for_impl().0;
 
     let (_, typ_generics, where_clause) = inp.generics.split_for_impl();