+++ minicbor-derive/src/encode.rs
@@ -48,8 +48,8 @@ fn on_struct(inp: &mut syn::DeriveInput) -> syn::Result<proc_macro2::TokenStream
         add_bound_to_type_params(bound, params, &blacklist, fields.fields().attributes(), Mode::Encode);
     }
 
-    let gen = add_typeparam(&inp.generics, gen_ctx_param()?, attrs.context_bound());
-    let impl_generics = gen.split_for_impl().0;
+    let generics = add_typeparam(&inp.generics, gen_ctx_param()?, attrs.context_bound());
+    let impl_generics = generics.split_for_impl().0;
 
     let (_, typ_generics, where_clause) = inp.generics.split_for_impl();
 
@@ -220,8 +220,8 @@ fn on_enum(inp: &mut syn::DeriveInput) -> syn::Result<proc_macro2::TokenStream> 
         add_bound_to_type_params(bound, params, &blacklist, &field_attrs, Mode::Encode);
     }
 
-    let gen = add_typeparam(&inp.generics, gen_ctx_param()?, enum_attrs.context_bound());
-    let impl_generics = gen.split_for_impl().0;
+    let generics = add_typeparam(&inp.generics, gen_ctx_param()?, enum_attrs.context_bound());
+    let impl_generics = generics.split_for_impl().0;
 
     let (_, typ_generics, where_clause) = inp.generics.split_for_impl();